[AN] refactor: 검색 화면 패키지 구조 변경 및 컴포넌트 분리 #716
*   [AN] refactor: 검색 메인 화면 관련 클래스 main 패키지로 이동
*   [AN] refactor: SearchMainScreen의 TopBar, CategoryList를 컴포저블로 분리
