+++ android/app/src/main/res/layout/layout_search_when_no_hearits.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="uiState"
-            type="com.onair.hearit.presentation.search.SearchUiState" />
+            type="com.onair.hearit.presentation.search.main.SearchUiState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
